% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Loss_of_CO2_fixing_pot.R
\name{Bog_plant_restoration}
\alias{Bog_plant_restoration}
\title{Bog_plant_restoration}
\usage{
Bog_plant_restoration(core.dat, AV_indirect)
}
\arguments{
\item{core.dat}{UI data}

\item{AV_indirect}{Area of drained peat}
}
\value{
L_fix
}
\description{
Bog_plant_restoration
}
