% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Loss_of_CO2_fixing_pot.R
\name{Bog_plant_sequestration_RM}
\alias{Bog_plant_sequestration_RM}
\title{Bog_plant_sequestration_RM}
\usage{
Bog_plant_sequestration_RM(core.dat, forestry.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{forestry.dat}{UI forestry data}
}
\value{
S_bog_plants
}
\description{
Bog_plant_sequestration_RM
}
