% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_Carbon_payback_time_peatland_restoration.R
\name{Carbon_payback_time}
\alias{Carbon_payback_time}
\title{Carbon_payback_time}
\usage{
Carbon_payback_time(res, sum_areas = T)
}
\arguments{
\item{res}{LCA output}
}
\value{
Carbon payback time/flux intercept
}
\description{
Carbon_payback_time
}
