% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carbon_mitigation_time_estimate.R
\name{CarbonMitigationMod}
\alias{CarbonMitigationMod}
\title{CarbonMitigationMod}
\usage{
CarbonMitigationMod(res, sum_areas = T)
}
\arguments{
\item{res}{LCA output}
}
\value{
Carbon payback time/flux intercept
}
\description{
CarbonMitigationMod
}
