% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_CO2_loss_by_DOC_and_POC_loss.R
\name{CO2_loss_DOC_POC_RM}
\alias{CO2_loss_DOC_POC_RM}
\title{CO2_loss_DOC_POC}
\usage{
CO2_loss_DOC_POC_RM(forestry.dat, L_gaseous, forest_soils = T)
}
\arguments{
\item{forestry.dat}{UI forestry data}

\item{L_gaseous}{Emissions object (forest soils or peatland)}

\item{forest_soils}{Select input from forest soils model}
}
\value{
L_DPOC
}
\description{
CO2_loss_DOC_POC
}
