% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5d_CO2_loss_from_drained_peat.R
\name{CO2_loss_drained}
\alias{CO2_loss_drained}
\title{CO2_loss_drained}
\usage{
CO2_loss_drained(core.dat, AV_indirect, R_tot)
}
\arguments{
\item{core.dat}{UI data}

\item{AV_indirect}{Area/Volume of drained peat}

\item{R_tot}{estimated emissions rates}
}
\value{
L_indirect
}
\description{
CO2_loss_drained
}
