% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5f_CO2_loss_from_restored_peat.R
\name{CO2_loss_rest_phase}
\alias{CO2_loss_rest_phase}
\title{CO2_loss_rest_phase}
\usage{
CO2_loss_rest_phase(core.dat, AV_indirect, R_tot)
}
\arguments{
\item{core.dat}{UI data}

\item{AV_indirect}{Area/Volume of drained peat}

\item{R_tot}{estimated emissions rates}
}
\value{
L_indirect
}
\description{
CO2_loss_rest_phase
}
