% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5f_CO2_loss_from_restored_peat.R
\name{CO2_loss_restoration}
\alias{CO2_loss_restoration}
\title{CO2_loss_restoration}
\usage{
CO2_loss_restoration(R_tot)
}
\arguments{
\item{R_tot}{estimated emissions rates}
}
\value{
L_indirect
}
\description{
CO2_loss_restoration
}
