% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{ECOSSR_CH4_AB}
\alias{ECOSSR_CH4_AB}
\title{ECOSSR_CH4_AB}
\usage{
ECOSSR_CH4_AB(CH4_CO2, d_wt, T_air)
}
\arguments{
\item{CH4_CO2}{CH4 to CO2 conversion factor}

\item{d_wt}{Average water table depth (undrained)}

\item{T_air}{Average air temperature}
}
\value{
CH4 emissions rate ECOSSE Acid bog
}
\description{
ECOSSR_CH4_AB
}
