% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{ECOSSR_CO2_F}
\alias{ECOSSR_CO2_F}
\title{ECOSSR_CO2_F}
\usage{
ECOSSR_CO2_F(CO2_C, d_wt, T_air)
}
\arguments{
\item{CO2_C}{Molecular weight ratio C to CO2}

\item{d_wt}{Average water table depth (undrained)}

\item{T_air}{Average air temperature}
}
\value{
CO2 emissions rate ECOSSE Fen
}
\description{
ECOSSR_CO2_F
}
