% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{Emissions_rates_forestry_soils_RM}
\alias{Emissions_rates_forestry_soils_RM}
\title{Emissions_rates_forestry_soils_RM}
\usage{
Emissions_rates_forestry_soils_RM(core.dat, forestry.dat, growthYield.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{forestry.dat}{UI forestry data}

\item{growthYield.dat}{growth and yield data (estimated from CARBINE runs)}
}
\value{
Emissions rates from drained soils under trees
}
\description{
Emissions_rates_forestry_soils_RM
}
