% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{Emissions_rates_soils_RM}
\alias{Emissions_rates_soils_RM}
\title{Emissions_rates_soils}
\usage{
Emissions_rates_soils_RM(core.dat, construct.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{construct.dat}{UI construction data}
}
\value{
Emissions rates from drained and undrained peatland
}
\description{
Emissions_rates_soils
}
