% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7ii_Forestry_CO2_loss_detail.R
\name{Forestry_CO2_loss_detail_RM}
\alias{Forestry_CO2_loss_detail_RM}
\title{Forestry_CO2_loss_detail}
\usage{
Forestry_CO2_loss_detail_RM(core.dat, forestry.dat, growthYield.dat, S_forest)
}
\arguments{
\item{core.dat}{UI data}

\item{forestry.dat}{UI forestry data}

\item{growthYield.dat}{growth and yield data (estimated from CARBINE runs)}

\item{C_wpry}{carbon content of the removed forestry estimated by 3PG}
}
\value{
Estimated lifetime loss of carbon stored in forestry products
}
\description{
Forestry_CO2_loss_detail
}
