% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{ForestSoilsEmissionsMod}
\alias{ForestSoilsEmissionsMod}
\title{ForestSoilsEmissionsMod}
\usage{
ForestSoilsEmissionsMod(forestry.dat, growthYield.dat)
}
\arguments{
\item{forestry.dat}{UI forestry data}

\item{growthYield.dat}{growth and yield data (estimated from CARBINE runs)}
}
\value{
Emissions rates from drained soils under trees
}
\description{
ForestSoilsEmissionsMod
}
