% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getCarbonDf}
\alias{getCarbonDf}
\title{getCarbonDf}
\usage{
getCarbonDf(
  S_forest,
  L_forest_soils,
  L_AqC_forest_soils,
  L_forest,
  L_peatland,
  L_AqC_peatland
)
}
\arguments{
\item{S_forest}{Forest sequestration}

\item{L_forest_soils}{Emissions from soils under forestry}

\item{L_AqC_forest_soils}{Aquatic carbon losses from forest soils}

\item{L_forest}{Forest product losses}

\item{L_peatland}{Emissions from peatland}

\item{L_AqC_peatland}{Aquatic carbon losses from oeatland}
}
\value{
Total peatland restoration carbon accounting dataframe
}
\description{
getCarbonDf
}
