% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{getDf1}
\alias{getDf1}
\title{getDf1}
\usage{
getDf1(list_ob)
}
\arguments{
\item{list_ob}{output of intermediate calculations in nested list format (Area, Exp/Min/Max)}
}
\value{
dataframe
}
\description{
getDf1
}
