% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{getDf2}
\alias{getDf2}
\title{getDf2}
\usage{
getDf2(list_ob)
}
\arguments{
\item{list_ob}{output of intermediate calculations in nested list format (Area, Exp/Min/Max)}
}
\value{
dataframe
}
\description{
getDf2
}
