% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{getDf3}
\alias{getDf3}
\title{getDf3}
\usage{
getDf3(list_ob)
}
\arguments{
\item{list_ob}{output of intermediate calculations in nested list format (Area, Exp/Min/Max)}
}
\value{
dataframe
}
\description{
getDf3
}
