% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_data.R
\name{getForestryInputData}
\alias{getForestryInputData}
\title{getForestryInputData}
\usage{
getForestryInputData(path, df_struct)
}
\arguments{
\item{path}{path to UI}

\item{df_struct}{dataframe summarising the data structure to output}
}
\value{
Forestry Input Data
}
\description{
getForestryInputData
}
