% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_data.R
\name{getYC}
\alias{getYC}
\title{getYC}
\usage{
getYC(input.dat, growthYield.dat)
}
\arguments{
\item{input.dat}{UI forestry data}

\item{growthYield.dat}{Growth and yield table}
}
\value{
Yield class estimated from UI average height/age
}
\description{
getYC
}
