% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Havesting_management_model.R
\name{HarvestingManagementMod}
\alias{HarvestingManagementMod}
\title{HarvestingManagementMod}
\usage{
HarvestingManagementMod(input.dat, S_forest)
}
\arguments{
\item{input.dat}{UI forestry data}

\item{S_forest}{3PG output}
}
\value{
Estimated lifetime loss of carbon stored in forestry products
}
\description{
HarvestingManagementMod
}
