% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{IPCC_CH4_AB}
\alias{IPCC_CH4_AB}
\title{IPCC_CH4_AB}
\usage{
IPCC_CH4_AB(CH4_CO2)
}
\arguments{
\item{CH4_CO2}{CH4 to CO2 conversion factor}
}
\value{
CH4 emissions rate IPCC Acid bog
}
\description{
IPCC_CH4_AB
}
