% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{IPCC_CH4_F}
\alias{IPCC_CH4_F}
\title{IPCC_CH4_F}
\usage{
IPCC_CH4_F(CH4_CO2)
}
\arguments{
\item{CH4_CO2}{CH4 to CO2 conversion factor}
}
\value{
CH4 emissions rate IPCC Fen
}
\description{
IPCC_CH4_F
}
