% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{list_op}
\alias{list_op}
\title{list_op}
\usage{
list_op(l1, l2, l3 = NULL, func)
}
\arguments{
\item{l1}{list object 1}

\item{l2}{list_object 2}

\item{l3}{list_object 3}

\item{func}{string describing elementwise operation}
}
\value{
outcome of elementwise operation
}
\description{
list_op
}
