% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Loss_of_CO2_fixing_pot.R
\name{Loss_of_CO2_fix_pot}
\alias{Loss_of_CO2_fix_pot}
\title{Loss_of_CO2_fix_pot}
\usage{
Loss_of_CO2_fix_pot(core.dat, AV_direct, AV_indirect)
}
\arguments{
\item{core.dat}{UI data}

\item{AV_direct}{Area of removed peat}

\item{AV_indirect}{Area of drained peat}
}
\value{
L_fix
}
\description{
Loss_of_CO2_fix_pot
}
