% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{METAR21_CH4}
\alias{METAR21_CH4}
\title{METAR21_CH4}
\usage{
METAR21_CH4(CO2_C, d_wt)
}
\arguments{
\item{d_wt}{Average water table depth (undrained)}

\item{CH4_CO2}{CH4 to CO2 conversion factor}
}
\value{
CH4 emissions rate meta-analytic regression Acid bog
}
\description{
METAR21_CH4
}
