% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{METAR21_CO2}
\alias{METAR21_CO2}
\title{METAR21_CO2}
\usage{
METAR21_CO2(CO2_C, d_wt)
}
\arguments{
\item{CO2_C}{Molecular weight ratio C to CO2}

\item{d_wt}{Average water table depth (undrained)}
}
\value{
CO2 emissions rate meta-analytic regression Acid bog
}
\description{
METAR21_CO2
}
