% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{METAR97_CO2_AB}
\alias{METAR97_CO2_AB}
\title{METAR97_CO2_AB}
\usage{
METAR97_CO2_AB(CO2_C, d_wt, T_air)
}
\arguments{
\item{CO2_C}{Molecular weight ratio C to CO2}

\item{d_wt}{Average water table depth (undrained)}

\item{T_air}{Average air temperature}
}
\value{
CO2 emissions rate meta-analytic regression Acid bog
}
\description{
METAR97_CO2_AB
}
