% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{METAR_CO2_AB}
\alias{METAR_CO2_AB}
\title{METAR_CO2_AB}
\usage{
METAR_CO2_AB(CO2_C, d_wt, T_air)
}
\arguments{
\item{CO2_C}{Molecular weight ratio C to CO2}

\item{d_wt}{Average water table depth (undrained)}

\item{T_air}{Average air temperature}
}
\value{
CO2 emissions rate meta-analytic regression Acid bog
}
\description{
METAR_CO2_AB
}
