% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Windfarm_CO2_emission_saving.R
\name{p_cap_windspeed}
\alias{p_cap_windspeed}
\title{p_cap_windspeed}
\usage{
p_cap_windspeed(core.dat, forestry.dat, R_windspeed_all)
}
\arguments{
\item{core.dat}{UI data}

\item{forestry.dat}{UI forestry data}

\item{R_windspeed_all}{estimated windspeed ratios}
}
\value{
power capacity
}
\description{
p_cap_windspeed
}
