% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{PeatlandSoilsEmissionsMod}
\alias{PeatlandSoilsEmissionsMod}
\title{PeatlandSoilsEmissionsMod}
\usage{
PeatlandSoilsEmissionsMod(input.dat)
}
\arguments{
\item{input.dat}{UI forestry data}

\item{R_tot}{estimated emissions rates}
}
\value{
L_indirect
}
\description{
PeatlandSoilsEmissionsMod
}
