% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotCounterfactual}
\alias{plotCounterfactual}
\title{plotCounterfactual}
\usage{
plotCounterfactual(res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotCounterfactual
}
