% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotES}
\alias{plotES}
\title{plotES}
\usage{
plotES(res, t_payback_res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{t_payback_res}{carbon payback estimate}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotES
}
