% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotL_AqC_forest_soils}
\alias{plotL_AqC_forest_soils}
\title{plotL_AqC_forest_soils}
\usage{
plotL_AqC_forest_soils(res)
}
\arguments{
\item{res}{LCA output}
}
\value{
plot
}
\description{
plotL_AqC_forest_soils
}
