% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotL_AqC_peatland}
\alias{plotL_AqC_peatland}
\title{plotL_AqC_peatland}
\usage{
plotL_AqC_peatland(res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotL_AqC_peatland
}
