% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotL_forest}
\alias{plotL_forest}
\title{plotL_forest}
\usage{
plotL_forest(res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotL_forest
}
