% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotL_forest_soils}
\alias{plotL_forest_soils}
\title{plotR_tot_forestry}
\usage{
plotL_forest_soils(L_forest_soils)
}
\arguments{
\item{plotL_forest_soils}{Emissions rates from soil under trees}
}
\value{
plot
}
\description{
plotR_tot_forestry
}
