% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotL_microbes}
\alias{plotL_microbes}
\title{plotL_microbes}
\usage{
plotL_microbes(L_microbes)
}
\arguments{
\item{L_microbes}{Microbial emissions output}
}
\value{
plot
}
\description{
plotL_microbes
}
