% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotL_peatland}
\alias{plotL_peatland}
\title{plotL_peatland}
\usage{
plotL_peatland(res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotL_peatland
}
