% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotLCA}
\alias{plotLCA}
\title{plotLCA}
\usage{
plotLCA(res, t_payback_res, sum_areas = T, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{t_payback_res}{carbon flux intercept estimate}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotLCA
}
