% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotR_tot_forestry}
\alias{plotR_tot_forestry}
\title{plotR_tot_forestry}
\usage{
plotR_tot_forestry(R_tot_forestry)
}
\arguments{
\item{R_tot_forestry}{Emissions rates from soil under trees}
}
\value{
plot
}
\description{
plotR_tot_forestry
}
