% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotS_bog_plants}
\alias{plotS_bog_plants}
\title{plotS_bog_plants}
\usage{
plotS_bog_plants(S_bog_plants)
}
\arguments{
\item{S_bog_plants}{BOg plant sequestration output}
}
\value{
plot
}
\description{
plotS_bog_plants
}
