% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotS_forest}
\alias{plotS_forest}
\title{plotS_forest}
\usage{
plotS_forest(res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotS_forest
}
