% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rest_dyn_mod}
\alias{rest_dyn_mod}
\title{rest_dyn_mod flexible convergent function for modelling restoration dynmaics}
\usage{
rest_dyn_mod(t, n, ymin, ymax, convThresh = 0.99)
}
\arguments{
\item{t}{sequence of years for restoration phase}

\item{n}{shape parameter (user input)}

\item{ymin}{value of y at t=0 (drained rate)}

\item{ymax}{value of y at t=max(t) (restored rate)}

\item{convThresh}{proportion of ymax at which convergence assumed to occur (since y attains ymax at t=Inf)}
}
\value{
rate sequence
}
\description{
rest_dyn_mod flexible convergent function for modelling restoration dynmaics
}
