% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aquatic_carbon_model.R
\name{AquaticCarbonMod}
\alias{AquaticCarbonMod}
\title{AquaticCarbonMod}
\usage{
AquaticCarbonMod(forestry.dat, L_gaseous, forest_soils = T)
}
\arguments{
\item{forestry.dat}{UI forestry data}

\item{L_gaseous}{Emissions object (forest soils or peatland)}

\item{forest_soils}{Select input from forest soils model}
}
\value{
L_DPOC
}
\description{
AquaticCarbonMod
}
