% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7a_C_sequest_in_trees_3PG.R
\name{C_sequest_in_trees_RM}
\alias{C_sequest_in_trees_RM}
\title{C_sequest_in_trees_RM}
\usage{
C_sequest_in_trees_RM(core.dat, forestry.dat)
}
\arguments{
\item{core.dat}{UI data}

\item{forestry.dat}{UI forestry data}
}
\value{
3PG output
}
\description{
C_sequest_in_trees_RM
}
