% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_Carbon_payback_time_peatland_restoration.R
\name{Carbon_payback_time}
\alias{Carbon_payback_time}
\title{Carbon_payback_time}
\usage{
Carbon_payback_time(
  S_forest,
  R_tot_forestry,
  S_bog_plants,
  L_forest,
  L_microbes,
  L_DPOC
)
}
\arguments{
\item{energy_output}{output of Windfarm_output}

\item{E_mat}{counterfactuals in matrix form with IDs as rownames}
}
\value{
Total Windfarm CO2 emissions savings
}
\description{
Carbon_payback_time
}
