% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1b_Carbon_payback_time_peatland_restoration.R
\name{Carbon_payback_time}
\alias{Carbon_payback_time}
\title{Carbon_payback_time}
\usage{
Carbon_payback_time(S_forest, R_tot_forestry, L_forest, L_peatland, L_DPOC)
}
\arguments{
\item{S_forest}{Forest sequestration}

\item{R_tot_forestry}{Emissions from soils under forestry}

\item{L_forest}{Forest product losses}

\item{L_peatland}{Emissions from peatland}

\item{L_DPOC}{D/POC losses}
}
\value{
Total peatland restoration carbon accounting
}
\description{
Carbon_payback_time
}
