% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_CO2_loss_by_DOC_and_POC_loss.R
\name{CO2_loss_DOC_POC_RM}
\alias{CO2_loss_DOC_POC_RM}
\title{CO2_loss_DOC_POC}
\usage{
CO2_loss_DOC_POC_RM(core.dat, L_peatland)
}
\arguments{
\item{core.dat}{UI data}

\item{L_peatland}{Peatland emissions object}
}
\value{
L_DPOC
}
\description{
CO2_loss_DOC_POC
}
