% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5e_Emission_rates_from_soils.R
\name{Emissions_rates_soils_RM}
\alias{Emissions_rates_soils_RM}
\title{Emissions_rates_soils}
\usage{
Emissions_rates_soils_RM(forestry.dat)
}
\arguments{
\item{forestry.dat}{UI forestry data}
}
\value{
Emissions rates from drained and undrained peatland
}
\description{
Emissions_rates_soils
}
