% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simplified_3PG.R
\name{ForestSequestrationMod}
\alias{ForestSequestrationMod}
\title{ForestSequestrationMod}
\usage{
ForestSequestrationMod(forestry.dat)
}
\arguments{
\item{forestry.dat}{UI forestry data}
}
\value{
3PG output
}
\description{
ForestSequestrationMod
}
