% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peatland_emissions_model.R
\name{ForestSoilsEmissionsMod}
\alias{ForestSoilsEmissionsMod}
\title{ForestSoilsEmissionsMod}
\usage{
ForestSoilsEmissionsMod(forestry.dat)
}
\arguments{
\item{forestry.dat}{UI forestry data}
}
\value{
Emissions rates from drained soils under trees
}
\description{
ForestSoilsEmissionsMod
}
