% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7ii_Forestry_CO2_loss_detail.R
\name{getYC}
\alias{getYC}
\title{getYC}
\usage{
getYC(forestry.dat, growthYield.dat)
}
\arguments{
\item{forestry.dat}{UI forestry data}

\item{growthYield.dat}{Growth and yield table}
}
\value{
Yield class estimated from UI average height/age
}
\description{
getYC
}
