% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Havesting_management_model.R
\name{HarvestingManagementMod}
\alias{HarvestingManagementMod}
\title{HarvestingManagementMod}
\usage{
HarvestingManagementMod(forestry.dat, growthYield.dat, S_forest)
}
\arguments{
\item{forestry.dat}{UI forestry data}

\item{growthYield.dat}{growth and yield data (estimated from CARBINE runs)}

\item{S_forest}{3PG output}
}
\value{
Estimated lifetime loss of carbon stored in forestry products
}
\description{
HarvestingManagementMod
}
