% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotL_AqC_forest_soils}
\alias{plotL_AqC_forest_soils}
\title{plotL_AqC_forest_soils}
\usage{
plotL_AqC_forest_soils(res, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotL_AqC_forest_soils
}
