% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotL_peatland}
\alias{plotL_peatland}
\title{plotL_peatland}
\usage{
plotL_peatland(L_peatland)
}
\arguments{
\item{L_peatland}{Microbial emissions output}
}
\value{
plot
}
\description{
plotL_peatland
}
