% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotLCA_cs}
\alias{plotLCA_cs}
\title{plotLCA}
\usage{
plotLCA_cs(res, t_payback_res, sum_areas = T, plabs = NULL)
}
\arguments{
\item{res}{LCA output}

\item{t_payback_res}{carbon payback estimate}

\item{plabs}{panel labels}
}
\value{
plot
}
\description{
plotLCA
}
